refactor: extract AppendTypeNameWithGenericArgs to shared TypeNameHelper
Addresses review feedback on #6075 — the extraction had duplicated the
method into both MetadataFilterMatcher and TestIdentifierService. Move it
to a single internal TUnit.Engine.Helpers.TypeNameHelper so the two call
sites cannot diverge.
